+++ .pre-commit-config.yaml
@@ -52,7 +52,7 @@ repos:
     hooks:
       - id: docs
         name: Docs
-        entry: poetry run ds_generate_docs docs
+        entry: poetry run generate_docs docs
         pass_filenames: false
         language: system
         files: '\.py$'
