+++ pyproject.toml
@@ -42,15 +42,14 @@ packages = [{ include = "docling_core" }]
 
 [tool.poetry.scripts]
 validate = "docling_core.utils.validate:main"
-ds_generate_jsonschema = "docling_core.utils.ds_generate_jsonschema:main"
-ds_generate_docs = "docling_core.utils.ds_generate_docs:main"
+generate_jsonschema = "docling_core.utils.generate_jsonschema:main"
+generate_docs = "docling_core.utils.generate_docs:main"
 
 [tool.poetry.dependencies]
 python = "^3.9"
 jsonschema = "^4.16.0"
 pydantic = "^2.6.0"
 jsonref = "^1.1.0"
-json-schema-for-humans = "^1.0.0"
 tabulate = "^0.9.0"
 pandas = "^2.1.4"
 pillow = "^10.3.0"
@@ -118,7 +117,6 @@ module = [
     "jsondiff.*",
     "jsonref.*",
     "jsonschema.*",
-    "json_schema_for_humans.*",
     "requests.*",
     "tabulate.*",
     "yaml.*",
